+++ src/operate.h
@@ -55,6 +55,7 @@ public:
     [[nodiscard]] bool isEcSysModuleLoaded() const;
     [[nodiscard]] bool loadEcSysModule() const;
     [[nodiscard]] bool updateEcData() const;
+    void updateEcDataAsync() const;
     [[nodiscard]] bool doProbe() const;
     [[nodiscard]] std::string getEcVersion() const;
     [[nodiscard]] std::string getEcBuild() const;