+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 All notable changes to this project will be documented in this file.
 
 ## Unreleased
+- Fixed UI lags when resizing the window (issue https://github.com/dmitry-s93/MControlCenter/issues/97)
 - Updated Norwegian Bokmål (Thanks [msigurdsen](https://github.com/msigurdsen))
 - Added support for German language (Thanks [EchterAlsFake](https://github.com/EchterAlsFake))
 - Added support for Hungarian language (Thanks [BiRo96](https://github.com/BiRo96))